requirements.txt
Django>=3.0.2,<3.1.0
Pillow==7.0.0
django-filter==2.2.0


